# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subsidy": {
      "post": {
        "operationId": "SubsidyController_requestSubsidy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestSubsidyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subsidy requested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubsidyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      },
      "get": {
        "operationId": "SubsidyController_listSubsidies",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "programName",
            "required": false,
            "in": "query",
            "description": "Filter by program name (case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or after this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateTo",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or before this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountMin",
            "required": false,
            "in": "query",
            "description": "Minimum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountMax",
            "required": false,
            "in": "query",
            "description": "Maximum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by subsidy status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "DISBURSED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SubsidyResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
    "/subsidy/{id}": {
      "get": {
        "operationId": "SubsidyController_getSubsidy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidy retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubsidyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
    "/subsidy/{id}/evidence": {
      "post": {
        "operationId": "SubsidyController_uploadEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadSubsidyEvidenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evidence uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SubsidyEvidenceResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
    "/subsidy/{id}/approve": {
      "patch": {
        "operationId": "SubsidyController_approveSubsidy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidy approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubsidyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
    "/subsidy/{id}/disburse": {
      "patch": {
        "operationId": "SubsidyController_disburseSubsidy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidy disbursed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubsidyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "RequestSubsidyDto": {
        "type": "object",
        "properties": {
          "onChainClaimId": {
            "type": "number",
            "description": "On-chain claim id (uint256) returned by the contract"
          },
          "onChainTxHash": {
            "type": "string",
            "description": "On-chain transaction hash (0x-prefixed)"
          },
          "amount": {
            "type": "number",
            "description": "Requested subsidy amount (in fiat equivalent)"
          },
          "programsId": {
            "type": "string",
            "description": "Associated programs id (must exist if provided)"
          },
          "weatherEventId": {
            "type": "string",
            "description": "Weather event id that triggered this request, if any",
            "nullable": true
          },
          "metadataHash": {
            "type": "string",
            "description": "Keccak256 hash (0x-prefixed) of the off-chain claim metadata bundle",
            "example": "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
          }
        },
        "required": [
          "onChainClaimId",
          "onChainTxHash",
          "amount",
          "metadataHash"
        ]
      },
      "FarmerDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nric": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "nric"
        ]
      },
      "SubsidyEvidenceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subsidyId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PHOTO",
              "PDF"
            ]
          },
          "storageUrl": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "description": "Original file name",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the uploaded file",
            "nullable": true
          },
          "fileSize": {
            "type": "number",
            "description": "Size of the uploaded file in bytes",
            "nullable": true
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subsidyId",
          "type",
          "storageUrl",
          "uploadedAt"
        ]
      },
      "SubsidyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "farmerId": {
            "type": "string"
          },
          "farmer": {
            "$ref": "#/components/schemas/FarmerDetailsDto"
          },
          "programsId": {
            "type": "string",
            "description": "Program id that triggered this request, if any",
            "nullable": true
          },
          "programName": {
            "type": "string",
            "description": "Program name",
            "nullable": true
          },
          "weatherEventId": {
            "type": "string",
            "description": "Weather event id that triggered this request, if any",
            "nullable": true
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "DISBURSED"
            ]
          },
          "metadataHash": {
            "type": "string",
            "description": "Keccak256 hash of off-chain claim payload",
            "nullable": true
          },
          "onChainClaimId": {
            "type": "number",
            "description": "On-chain claim id (as string for bigint safety)"
          },
          "onChainTxHash": {
            "type": "string",
            "description": "Tx hash for on-chain interactions",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "description": "If rejected, the reason",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "approvedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time the subsidy was approved",
            "nullable": true
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time the subsidy was paid",
            "nullable": true
          },
          "evidences": {
            "description": "Evidence documents uploaded for this subsidy",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubsidyEvidenceResponseDto"
            }
          }
        },
        "required": [
          "id",
          "farmerId",
          "amount",
          "status",
          "createdAt"
        ]
      },
...
      "UploadSubsidyEvidenceDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Photo or PDF evidence for the subsidy claim"
          }
        },
        "required": [
          "file"
        ]
      },
...
    }
  }
}